LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-interpolation-expected.txt

CONSOLE MESSAGE: Error: assert_true: 'underlying' value must be supported expected true got false

Harness Error (FAIL), message = Error: assert_true: 'underlying' value must be supported expected true got false
